Fix fp16 checkpoints crashing the head, quiet HF progress bars, auto-submit
mDeBERTa-v3 died on Kaggle with "mat1 and mat2 must have the same dtype,
but got Half and Float" inside the classifier head. Its published weights
are stored in fp16 and transformers keeps the checkpoint's dtype, so the
backbone emitted Half while the head was Float. configs/deberta.yaml even
asked for precision: fp32, but that flag only governs autocast, never the
weights. Both branches of TransformerClassifier now pin the backbone to
fp32 - AMP wants fp32 master weights regardless - including the from_config
branch, where a config saved off an fp16 model rebuilt an fp16 backbone and
load_state_dict copied into it in place, so checkpoint inference failed the
same way.

Hugging Face draws tqdm bars while downloading and materialising weights.
Off a TTY, one bar becomes hundreds of "Loading weights: 3%|..." lines.
factory.py disables them (HASTIKA_HF_VERBOSE=1 brings them back) and
train.py logs the load itself so the silence is not mistaken for a hang.

train.py now writes a ready-to-submit predictions.csv + submission.zip for
the organisers' validation file at the end of every run, named
{task}_val_{run}, and for the test file once that exists. The writer moved
to src/evaluation/submission.py, shared with inference.py. Checked against
docs/starting_kit: same header, label spelling, row count and id order.

+++ src/models/classifier.py
@@ -13,9 +13,15 @@ class TransformerClassifier(nn.Module):
                  dropout: float = 0.1, backbone_config=None):
         super().__init__()
         if backbone_config is None:                      # training: load pretrained weights
-            self.backbone = AutoModel.from_pretrained(backbone_name)
+            # .float() is not redundant: some published checkpoints store fp16 weights
+            # (mDeBERTa-v3 does) and transformers keeps the checkpoint's dtype, which then
+            # meets the fp32 head as "mat1 and mat2 must have the same dtype". AMP wants fp32
+            # master weights anyway -- mixed precision is applied by autocast, not by the weights.
+            self.backbone = AutoModel.from_pretrained(backbone_name).float()
         else:                                            # inference: architecture only, weights come from ckpt
-            self.backbone = AutoModel.from_config(backbone_config)
+            # .float() for the same reason, from the other direction: a config saved off an fp16
+            # backbone builds an fp16 one, and load_state_dict copies in place, so fp16 would stick.
+            self.backbone = AutoModel.from_config(backbone_config).float()
         self.meta = {"backbone_name": backbone_name, "num_labels": num_labels,
                      "pooling": pooling, "dropout": dropout}
         self.pooling = pooling
